feat: fetch and display country names from API
This commit refactors a React functional component to fetch and
display a list of country names from the Rest Countries API.
The useState hook is used to define a state variable 'countries' and
its corresponding setter function 'setCountries', with an initial value
of 'null'. An asynchronous 'fetcher' function is defined using the axios
library to fetch data from the API. An asynchronous 'fetchData' function
is defined that uses 'fetcher' to fetch data from the API and sets the
'countries' state variable to the fetched data. The 'fetchData' function
is called when the component is mounted. The useState hook is called with
a callback function that executes the 'fetchData' function only once when
the component is mounted.

In the component's return statement, an unordered list is displayed with
each country name displayed as a list item, with the 'countries' state
variable checked to ensure that it is not 'null' before mapping over it
to display the list items. The 'c.name.common' property is accessed to
display the name of each country in the list.