feat: add region filter select to HomePage
This commit adds a select dropdown menu to filter countries by region. The `select` element is added to the `header` section along with the title. The selected region state is managed using the `useCountryStore` hook. The available options in the `select` element are "All", "America", "Oceania", and "Asia".

The list of countries is filtered based on the selected region. The `handleRegionSelect` function is called whenever a region is selected from the dropdown.

This commit also updates the `handleCountryClick` function to set the `selectedCountry` state using the `setSelectedCountry` function from the `useCountryStore` hook.