feat: Add bordering countries to country page
This commit adds the ability to display a list of bordering
countries on the country page.
The bordering countries are displayed as clickable links
to the respective country pages.

The code for the country page has also been
updated for better readability.

pages/countries/[code].tsx:
- Added bordering countries feature to display a list of clickable
links to bordering countries on the country page.
- Updated the code for better readability.