feat: retrieve border countries as links
Display the border countries on each dynamic country page
in the application.

The function fetchBordersWithCountries retrieves the data for an
array of border codes from the Rest Countries API, and returns
an array of ICountry objects for those borders that are found.

It uses the fetcher utility function to get the data from the API,
and also retrieves all the countries from the API to be used as
a lookup table for border countries.