feat: handle immutable state updates with Immer
This commit adds the Immer library to the project for immutable state updates.
We are now using Immer's `produce` function to create a draft copy of the app state.
The function can directly modify this draft copy. Once the function has completed,
Immer compares the draft copy to the original state, and generates a new immutable state
based on the changes made to the draft. This allows us to write code that looks like
we're directly mutating the state, but in reality, Immer is doing the heavy lifting to
ensure that our state updates are handled correctly.

We also updated the `useCountryStore` function in the `country-store.ts` file to use
Immer's `produce` function for state updates instead of the spread operator. This makes
our state updates more concise and easier to read. Finally, we added a new import statement
to the `countries/[code].tsx` file to import the `useCountryStore` hook from the `country-store.ts`
file so that we can use the `setSelectedCountry` function to update the app state.

Also a go back button was added to use `useRouter` `.back` method, to go
back in hidtory.